// apps/desktop/src/ipc/generated/ColourGrading.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GradeWheel } from "./GradeWheel";

export type ColourGrading = { shadows: GradeWheel, midtones: GradeWheel, highlights: GradeWheel, global: GradeWheel, 
/**
 * How much the ranges overlap, 0..100 (50 as Lightroom's default).
 */
blending: number, 
/**
 * Where shadows give way to highlights, -100..100: positive gives highlights more
 * of the range.
 */
balance: number, };